LayoutTests/fast/css/getPropertyValue-returns-empty-on-missing-property-expected.txt

Verify that CSSStyleDeclaration.getPropertyValue() returns an empty string when the property argument does not match.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('target').style.getPropertyValue() threw exception TypeError: Not enough arguments.
PASS document.getElementById('target').style.getPropertyValue('foobar') is ""
PASS document.getElementById('target').style.getPropertyValue('display') is "none"
PASS document.getElementById('target').style.getPropertyValue('color') is "rgba(1, 2, 3, 0.4)"
PASS document.getElementById('target').style.getPropertyValue('-webkit-display') is ""
PASS document.getElementById('target').style.getPropertyValue('-webkit-color') is ""
PASS document.getElementById('target').style.getPropertyValue('foobar', 1) is ""
PASS document.getElementById('target').style.getPropertyValue('display', null) is "none"
PASS document.getElementById('target').style.getPropertyValue('color', undefined) is "rgba(1, 2, 3, 0.4)"
PASS document.getElementById('target').style.getPropertyValue('-webkit-display', 'unused') is ""
PASS document.getElementById('target').style.getPropertyValue('-webkit-color', { foo:'bar'}) is ""
PASS successfullyParsed is true

TEST COMPLETE
